refactor: modularize CLI command structure
- Split `src/commands/content.js` into `src/commands/courses.js`, `src/commands/modules.js`, and `src/commands/sections.js`.
- Export resource-specific setup functions: `setupCoursesCommands`, `setupModulesCommands`, and `setupSectionsCommands`.
- Update `src/index.js` to initialize new command modules and remove `setupContentCommands`.
- Maintain full feature parity and user-facing command syntax.
- Cleanup unused `chalk` import in `src/index.js`.
